remove project contract compiler
